requirements.txt
Django==1.4
South==0.7.5
Werkzeug==0.8.3
argparse==1.2.1
django-annoying==0.7.6
django-debug-toolbar==0.9.4
django-extensions==0.9
django-model-utils==1.1.0
django-widget-tweaks==1.1.2
psycopg2==2.4.5
wsgiref==0.1.2
